+++ lite/client/main_test.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/tendermint/abci/example/dummy"
+	"github.com/tendermint/abci/example/kvstore"
 
 	nm "github.com/tendermint/tendermint/node"
 	rpctest "github.com/tendermint/tendermint/rpc/test"
@@ -14,7 +14,7 @@ var node *nm.Node
 
 func TestMain(m *testing.M) {
 	// start a tendermint node (and merkleeyes) in the background to test against
-	app := dummy.NewDummyApplication()
+	app := kvstore.NewKVStoreApplication()
 	node = rpctest.StartTendermint(app)
 	code := m.Run()