+++ test/p2p/test.sh
@@ -4,7 +4,7 @@ set -eu
 DOCKER_IMAGE=$1
 NETWORK_NAME=local_testnet
 N=4
-PROXY_APP=persistent_dummy
+PROXY_APP=persistent_kvstore
 
 cd "$GOPATH/src/github.com/tendermint/tendermint"
 
@@ -13,7 +13,7 @@ set +e
 bash test/p2p/local_testnet_stop.sh "$NETWORK_NAME" "$N"
 set -e
 
-PERSISTENT_PEERS=$(bash test/p2p/persistent_peers.sh $N)
+PERSISTENT_PEERS=$(bash test/p2p/persistent_peers.sh $N $DOCKER_IMAGE)
 
 # start the testnet on a local network
 # NOTE we re-use the same network for all tests